refactor(kanban): extract txn/task-link mixins from kanban_db.py (shard s2)
