+++ hermes_cli/kanban_db_events.py
@@ -1,0 +1,191 @@
+"""Kanban task events and run lifecycle: event rows, run close, and
+synthetic (never-claimed) run rows.
+
+Extracted verbatim from :mod:`hermes_cli.kanban_db` (godfile shard s2,
+cluster c11). Public names are re-exported from :mod:`hermes_cli.kanban_db`
+so existing ``import hermes_cli.kanban_db as kb`` callers are unaffected.
+"""
+
+from __future__ import annotations
+
+import json
+import sqlite3
+import time
+from typing import Optional
+
+# ---------------------------------------------------------------------------
+# Verbatim section moved from hermes_cli/kanban_db.py (lines 3930-4090).
+# The kanban_db.py -> sibling re-export cycle is import-safe because the
+# parent-module imports at the bottom of this file are deferred to the end
+# of the module (see kanban_db.py bottom re-export block).
+# ---------------------------------------------------------------------------
+
+def list_events(conn: sqlite3.Connection, task_id: str) -> list[Event]:
+    rows = conn.execute(
+        "SELECT * FROM task_events WHERE task_id = ? ORDER BY created_at ASC, id ASC",
+        (task_id,),
+    ).fetchall()
+    out = []
+    for r in rows:
+        try:
+            payload = json.loads(r["payload"]) if r["payload"] else None
+        except Exception:
+            payload = None
+        out.append(
+            Event(
+                id=r["id"],
+                task_id=r["task_id"],
+                kind=r["kind"],
+                payload=payload,
+                created_at=r["created_at"],
+                run_id=(int(r["run_id"]) if "run_id" in r.keys() and r["run_id"] is not None else None),
+            )
+        )
+    return out
+
+
+def _append_event(
+    conn: sqlite3.Connection,
+    task_id: str,
+    kind: str,
+    payload: Optional[dict] = None,
+    *,
+    run_id: Optional[int] = None,
+) -> None:
+    """Record an event row.  Called from within an already-open txn.
+
+    ``run_id`` is optional: pass the current run id so UIs can group
+    events by attempt. For events that aren't scoped to a single run
+    (task created/edited/archived, dependency promotion) leave it None
+    and the row carries NULL.
+    """
+    now = int(time.time())
+    pl = json.dumps(payload, ensure_ascii=False) if payload else None
+    conn.execute(
+        "INSERT INTO task_events (task_id, run_id, kind, payload, created_at) "
+        "VALUES (?, ?, ?, ?, ?)",
+        (task_id, run_id, kind, pl, now),
+    )
+
+
+def _end_run(
+    conn: sqlite3.Connection,
+    task_id: str,
+    *,
+    outcome: str,
+    summary: Optional[str] = None,
+    error: Optional[str] = None,
+    metadata: Optional[dict] = None,
+    status: Optional[str] = None,
+) -> Optional[int]:
+    """Close the currently-active run for ``task_id`` and clear the pointer.
+
+    ``outcome`` is the semantic result (completed / blocked / crashed /
+    timed_out / spawn_failed / gave_up / reclaimed). ``status`` is the
+    run-row status (usually just ``outcome``, but callers can pass it
+    explicitly). Returns the closed run_id or ``None`` if no active run
+    existed (e.g. a CLI user calling ``hermes kanban complete`` on a
+    task that was never claimed).
+    """
+    now = int(time.time())
+    row = conn.execute(
+        "SELECT current_run_id FROM tasks WHERE id = ?", (task_id,),
+    ).fetchone()
+    if not row or not row["current_run_id"]:
+        return None
+    run_id = int(row["current_run_id"])
+    conn.execute(
+        """
+        UPDATE task_runs
+           SET status        = ?,
+               outcome       = ?,
+               summary       = ?,
+               error         = ?,
+               metadata      = ?,
+               ended_at      = ?,
+               claim_lock    = NULL,
+               claim_expires = NULL,
+               worker_pid    = NULL
+         WHERE id = ?
+           AND ended_at IS NULL
+        """,
+        (
+            status or outcome,
+            outcome,
+            summary,
+            error,
+            json.dumps(metadata, ensure_ascii=False) if metadata else None,
+            now,
+            run_id,
+        ),
+    )
+    conn.execute(
+        "UPDATE tasks SET current_run_id = NULL WHERE id = ?", (task_id,),
+    )
+    return run_id
+
+
+def _current_run_id(conn: sqlite3.Connection, task_id: str) -> Optional[int]:
+    row = conn.execute(
+        "SELECT current_run_id FROM tasks WHERE id = ?", (task_id,),
+    ).fetchone()
+    return int(row["current_run_id"]) if row and row["current_run_id"] else None
+
+
+def _synthesize_ended_run(
+    conn: sqlite3.Connection,
+    task_id: str,
+    *,
+    outcome: str,
+    summary: Optional[str] = None,
+    error: Optional[str] = None,
+    metadata: Optional[dict] = None,
+) -> int:
+    """Insert a zero-duration, already-closed run row.
+
+    Used when a terminal transition happens on a task that was never
+    claimed (CLI user calling ``hermes kanban complete <ready-task>
+    --summary X``, or dashboard "mark done" on a ready task). Without
+    this, the handoff fields (summary / metadata / error) would be
+    silently dropped: ``_end_run`` is a no-op because there's no
+    current run.
+
+    The synthetic run has ``started_at == ended_at == now`` so it
+    shows up in attempt history as "instant" and doesn't skew elapsed
+    stats. Caller is responsible for leaving ``current_run_id`` NULL
+    (or for clearing it elsewhere in the same txn) since this
+    function does NOT touch the tasks row.
+    """
+    now = int(time.time())
+    trow = conn.execute(
+        "SELECT assignee, current_step_key FROM tasks WHERE id = ?",
+        (task_id,),
+    ).fetchone()
+    profile = trow["assignee"] if trow else None
+    step_key = trow["current_step_key"] if trow else None
+    cur = conn.execute(
+        """
+        INSERT INTO task_runs (
+            task_id, profile, step_key,
+            status, outcome,
+            summary, error, metadata,
+            started_at, ended_at
+        ) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?)
+        """,
+        (
+            task_id, profile, step_key,
+            outcome, outcome,
+            summary, error,
+            json.dumps(metadata, ensure_ascii=False) if metadata else None,
+            now, now,
+        ),
+    )
+    return int(cur.lastrowid or 0)
+
+# Names from the parent module (module-level dataclasses that stay in
+# kanban_db.py because other clusters use them). Imported at the bottom of
+# this module so the kanban_db <-> sibling re-export cycle resolves in any
+# import order.
+from hermes_cli.kanban_db import (  # noqa: E402
+    Event,
+)
